requirements.txt
wheel
django
djangorestframework
pygments
psycopg2-binary
django-cors-headers
django-cors-middleware
python-decouple
dj-database-url
whitenoise
gunicorn